Refactor, remove duplicate code
