+++ src/pages/LendPage/components/PlaceOfferTab/PlaceOfferTab.module.less
@@ -49,3 +49,7 @@
     background: var(--accent-primary);
   }
 }
+
+.borrowerMessage {
+  font: var(--body-text-sm);
+}
