+++ src/pages/LendPage/components/PlaceOfferTab/helpers.ts
@@ -1,0 +1,18 @@
+type ShouldShowDepositError = (props: {
+  initialLoanValue?: number
+  initialLoansAmount?: number
+  solanaBalance: number
+  offerSize: number
+}) => boolean
+
+export const shouldShowDepositError: ShouldShowDepositError = ({
+  initialLoanValue = 0,
+  initialLoansAmount = 0,
+  solanaBalance,
+  offerSize,
+}) => {
+  const initialOfferSize = initialLoansAmount * initialLoanValue
+  const totalAvailableFunds = initialOfferSize + solanaBalance
+
+  return totalAvailableFunds < offerSize
+}
