+++ src/pages/LendPage/components/PlaceOfferTab/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './useOfferTransactions'
+export * from './usePlaceOffer'
